app/requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
gunicorn>=20.0.4,<20.1.0
django-cors-headers>=3.7.0,<3.8.0

django-health-check>=3.16.4,<3.17

flake8>=3.8.4,<3.9.0
psycopg2-binary>=2.8.6,<2.9.0

drf-yasg>=1.20.0,<1.21.0

celery>=5.1.2,<5.2.0
redis>=3.5.3,<3.6.0